// web/lib/api/generated/model/ledgerAdjustmentKindEnum.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `OPENING_DEBIT` - Opening balance (owed)
 * * `OPENING_CREDIT` - Opening balance (advance)
 * * `DEBIT` - Debit adjustment
 * * `CREDIT` - Credit adjustment
 */
export type LedgerAdjustmentKindEnum = typeof LedgerAdjustmentKindEnum[keyof typeof LedgerAdjustmentKindEnum];


export const LedgerAdjustmentKindEnum = {
  OPENING_DEBIT: 'OPENING_DEBIT',
  OPENING_CREDIT: 'OPENING_CREDIT',
  DEBIT: 'DEBIT',
  CREDIT: 'CREDIT',
} as const;