// web/lib/api/generated/model/receivablesAgeingBasis.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */

export type ReceivablesAgeingBasis = typeof ReceivablesAgeingBasis[keyof typeof ReceivablesAgeingBasis];


export const ReceivablesAgeingBasis = {
  DUE_DATE: 'DUE_DATE',
  INVOICE_DATE: 'INVOICE_DATE',
} as const;